Allow loser 3.6 test_reforms logic
